Continued in rewriting code to 64bit

+++ constable/hash.c
@@ -8,7 +8,7 @@
 
 #define hash_func(x)	((((x)>>24)^((x)>>16)^((x)>>8)^x)&0x00ff)
 
-void hash_add( struct hash_s *h, struct hash_ent_s *e, u_int32_t key )
+void hash_add( struct hash_s *h, struct hash_ent_s *e, uintptr_t key )
 { int i;
 	i=hash_func(key);
 	e->next=h->table[i];
@@ -16,7 +16,7 @@ void hash_add( struct hash_s *h, struct hash_ent_s *e, u_int32_t key )
 	h->table[i]=e;
 }
 
-struct hash_ent_s *hash_find( struct hash_s *h, u_int32_t key )
+struct hash_ent_s *hash_find( struct hash_s *h, uintptr_t key )
 { struct hash_ent_s *e;
 	for(e=h->table[hash_func(key)];e!=NULL;e=e->next)
 		if( e->key==key )
